+++ src/templates/gitignore.ts
@@ -1,0 +1,5 @@
+export const templateGitIgnore = `.gitignore
+entity.cmake
+entity.pri
+$GenerateFileList$
+`;